requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
lightgbm>=4.0
matplotlib>=3.7
seaborn>=0.12
holidays>=0.40
airportsdata>=20230101
google-cloud-bigquery
db-dtypes
